Update uroot.py
Big changes made to root calculation class.  I now have a base RootFind class and sub-class all the different methods.  Some methods are not so useful but included for completeness.  My favorite is 1D Broyden.  I've hacked this together from the conventional multi-dim Broyden method. It's not really recognized as an actual method in itself because it can be shown to resolve to Secant.  But, the structure has a single guess and only 1 function call per loop.  This makes it good for use in iterative calculations.  So far, I have found it to be reliable and fast.